PPP: fix default size of protocol field
In commit 2f5d9bd4dab8 ("PPP: protocol field can be limited to one byte"),
a new class PPP_ was added to manage the ability to parse and generate a
one byte PPP protocol field.

This was later reworked by commit 834309f91c1d ("Small doc cleanups"), which
removed the PPP_ class, and changed the default behavior of PPP to generate a
one byte protocol field by default, when < 0x100.

The RFC states that "by default, all implementations MUST transmit packets with
two octet PPP Protocol fields". The one byte field are issued by implementations
when the compression is negociated.

This patch reverts to the original behavior, which is to generate a 2 bytes
protocol field by default, but make it possible to explicitly generate a one
byte protocol by passing the value as bytes(). The PPP class is still able to
parse either a one or two bytes protocol.

Link: https://www.rfc-editor.org/rfc/rfc1661.html#section-6.5
Fixes: 834309f91c1d ("Small doc cleanups")
Fixes : 3913